pyproject.toml
[project]
name = "notion-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "pillow>=11.2.1",
    "pillow-heif>=0.22.0",
    "requests>=2.32.3",
]
